requirements-test.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.5
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
requests-mock==1.11.0
responses==0.24.1
